% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login_page.R
\name{login_page_ui}
\alias{login_page_ui}
\alias{login_page_server}
\title{mod_login_page_ui and mod_login_page_server}
\usage{
login_page_ui(id)

login_page_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
