% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clients_database.R
\name{clients_database_ui}
\alias{clients_database_ui}
\alias{clients_database_server}
\title{mod_clients_database_ui and mod_clients_database_server}
\usage{
clients_database_ui(id)

clients_database_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
