% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_register_form.R
\name{register_form_ui}
\alias{register_form_ui}
\alias{register_form_server}
\title{mod_register_form_ui and mod_register_form_server}
\usage{
register_form_ui(id)

register_form_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
