% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_owner.R
\docType{class}
\name{owner_data-class}
\alias{owner_data-class}
\title{An S4 class to represent an owner file.}
\description{
An S4 class to represent an owner file.
}
\section{Slots}{

\describe{
\item{\code{first_name}}{First Name of the owner}

\item{\code{last_name}}{Last Name of the owner}

\item{\code{street_name}}{Street Name of the owner's address}

\item{\code{post_code}}{Post Code of the owner's address}

\item{\code{email}}{email address the owner wishes to be contacted with}

\item{\code{phone_number}}{Phone Number the owner wises to be contacted from}
}}

