% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Carries over the list of arguments available in golem::with_golem_options}
}
\description{
See \code{shiny::\link[shiny:shinyApp]{shinyApp}} for details.
}
