% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar_ui}
\alias{sidebar_ui}
\alias{sidebar_server}
\title{Sidebar Module}
\usage{
sidebar_ui(id)

sidebar_server(input, output, session, log_status)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
