% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slotApply}
\alias{slotApply}
\title{Function that loops over the 'slots' of a given object. In the same fashion the 'Apply' function would work over the elements of a list.}
\usage{
slotApply(x, FUN, ...)
}
\arguments{
\item{x}{object to be looped upoon}

\item{FUN}{custom function to be applied upon the object}

\item{...}{further parameters to be carried over to the custom function}
}
\description{
Function that loops over the 'slots' of a given object. In the same fashion the 'Apply' function would work over the elements of a list.
}
