% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_owner.R
\name{updateData,owner_data-method}
\alias{updateData,owner_data-method}
\title{An S4 method to update the existing data of class 'owner_data'}
\usage{
\S4method{updateData}{owner_data}(owner_data)
}
\arguments{
\item{owner_data}{An S4 object of class 'owner_data'}
}
\description{
An S4 method to update the existing data of class 'owner_data'
}
