% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_owner.R
\name{updateData}
\alias{updateData}
\title{An S4 generic to update existing data of some specific class}
\usage{
updateData(owner_data)
}
\arguments{
\item{owner_data}{A data.frame, the existing file of Owners' Data}
}
\description{
An S4 generic to update existing data of some specific class
}
